Greatly simplified the 'StageRotation' class, by removing the restrictions that its quaternion may not represent the identity rotation, and its time delta may not be zero.  This simplification means that while StageRotations may be created from ANY unit quaternion and time delta (thus removing the need for the static 'CreateStageRotation' member function, which checked that the unit quaternion and time delta were ok, before finally invoking the protected constructor), not all StageRotations are automatically valid for scaling by time.
This scaling is now done by the function 'scaleToNewTimeDelta', which takes
a StageRotation and a new time delta, and does all the checks and calculations
(reverse-engineering of the rotation axis and rotation angle) itself.  Again,
this makes the StageRotation class much simpler, since it now doesn't contain
the rotation axis and rotation angle as members.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@289 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c