Changed the internal implementation of StringSet from a 'std::set' to an intrusive-ptr-to-StringSetImpl object, which in turn contains the 'std::set'.
This is a fairly major change.  Now, the StringSet is little more than a handle
which wraps the reference-counted StringSetImpl and provides an interface to
manipulate its contents.  'StringSet::SharedIterator' instances now contain
intrusive-ptr-to-StringSetImpl rather than ptr-to-'std::set', so the StringSet
shares the memory-management of the StringSetImpl with all the SharedIterators.

This whole change was motivated by the realisation that when StringSet
contained the 'std::set' directly, the destruction of the StringSet meant that
every non-default-constructed SharedIterator would now contain a (non-NULL)
pointer to a 'std::set' instance which no longer existed, as well as an
iterator pointing to an element of the 'std::set'; when a SharedIterator is
destructing, if its pointer is non-NULL, it attempts to decrement the
reference-count of the element through the iterator...  In a nutshell, if any
non-default-constructed SharedIterators were to outlive their StringSet, their
destruction would be undefined!

This problem is solved quite nicely by changing the ptr-to-'std::set' in the
SharedIterator to intrusive-ptr-to-StringSetImpl:  Even if the StringSet
instance is destroyed, as long as there are SharedIterator instances which
refer to it, the StringSetImpl will persist.

In summary, the changes include:
 * definition of 'StringSet::StringSetImpl';
 * changing collection-ptr to impl-ptr in 'StringSet::SharedIterator';
 * changing direct 'std::set' accesses to indirections through 'd_impl' in StringSet.
 * addition of the functions used by intrusive_ptr to increment/decrement refs.

The code compiles, but could probably do with closer inspection; I'm just
committing it tonight for safe-keeping.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@908 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c