The beginnings of a Visitor pattern implementation for GPML output.
Added to FeatureHandle the first elements of a (far from complete) Visitor
pattern structure for the recursive traversal of features, as well as the first
elements of a (far from complete) GPML-1.5 output visitor which makes use of
this Visitor pattern structure.

Currently, this output visitor is able to "traverse" FeatureHandle, but doesn't
attempt to recurse below this.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@932 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c