Implemented composition of rotations in the reconstruction-tree.  Also modified the reconstruction-tree-traversal functions to output the resulting "composed absolute rotation".
Superficially, it all looks correct, but I can't really tell until I can see
geometries moving on the screen.  The code-changes themselves are pretty
simple:  The new code is basically the same as the old code (which I also
wrote) in the function '::CheckRotation' in "src/controls/Reconstruct.cc"; 
I'm also made more confident by the (surprisingly) lucid comment in front of
the '::GPlatesMaths::compose' function itself, which provides a clear and
helpful specification of the correct order of operands.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@1081 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c