Re-enabled the building of the "gplates_new" executable in "src/Makefile.am".
It was disabled in changeset:1103 due to "compilation errors"; these
compilation errors were in fact caused by the build system attempting to link
"controls/libcontrols.a", "fileio/libfileio.a", "geo/libgeo.a" and
"global/libglobal.a" into the executable, after the subdirectories "controls",
"fileio", "geo" and "global" had been removed from the list of subdirectories
to build (in changeset:1016) on the
source:/gplates/branches/qt-gui-prototype-2006-nov-29#1103 branch.

The reason this problem hadn't been noticed before was that I hadn't done a
"clean build" on the
source:/gplates/branches/reconstructions-in-gpgim-2006-nov-23#1079 branch, so
the libraries were still present from previous builds (before the changes to
"src/Makefile.am" were ported to the branch in changeset:1068).


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@1114 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c