Modified FeatureHandle to use the templatised smart iterator type 'HandleContainerIterator', just as FeatureCollectionHandle and FeatureStoreRootHandle do, rather than providing direct access to the container of properties.
Several typedefs were added to FeatureHandle to enable the templatised iterator to traverse the contents of a FeatureHandle.  Additionally, the direct access (returning a reference to the member datum inside FeatureRevision) to the container of properties was replaced with: functions to return "begin" and "end" positions of the smart iterator; and the functions 'append_property_container' and 'remove_property_container'.

FeatureRevision was also modified to be more like FeatureCollectionRevision and FeatureStoreRootRevision, by replacing direct access to the container of properties with the functions: 'size'; 'operator[]'; 'access_property_container'; 'append_property_container'; and 'remove_property_container'.

Modified the FeatureHandle-construction code in "src/model/Model.cc" and "src/gplates_new_main.cc" to use the FeatureHandle member-function 'append_property_container' rather than the (now removed) 'properties' accessor.

Updated the feature visitors ReconstructionTreePopulator, ReconstructedFeatureGeometryPopulator and GpmlOnePointFiveOutputVisitor to use the smart iterators (const and non-const) rather than accessing the container of properties directly.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@1228 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c