Fixed an error in "src/gplates_new_main.cc" (a transaction wasn't committed).
Although the transactions are currently only "dummy" transactions (implemented
as class DummyTransactionHandle), and invoking "commit" is a no-op as far as
the transaction mechanism goes, we want the client code to be as correct as
possible at this stage, so that when the "dummy" transactions become real
transactions, everything works correctly.

This error was indicated as a result of the changes applied in changeset:1234.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@1235 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c