Replaced source:/gplates/trunk using source:/gplates/branches/plates-fileio-model-integration-2006-nov-21#1418 incorporating log:/gplates/branches/plates-fileio-model-integration-2006-nov-21#979:1418
This is "[wiki:MergingInSubversion#Replacingonebranchwithanother replacing one branch with another]". The command used was:

{{{
svn merge http://svn.gplates.org/gplates/trunk@1418 http://svn.gplates.org/gplates/branches/plates-fileio-model-integration-2006-nov-21@1418
}}}
from inside the root directory of a working copy of the trunk. 

The contents of the trunk were last synched to the branch in changeset:1350.  It may be observed that there haven't been any commits to the trunk between changeset 1308 and this changeset (see log:/gplates/trunk#1308:1419), so there was no need to sync the trunk to this branch before the merge.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@1419 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c