Modified class GeoTimeInstant to use a TimePositionType enumeration.
Previously, class GeoTimeInstant contained only a single member datum of type 'double'.  The time-positions for the distant past and the distant future were represented using the floating-point values for infinity and minus-infinity (respectively).

However, when Glen was porting to Windows (specifically, the Microsoft Visual Studio C++ compiler), he pointed out that: 

{{{
1. INFINITY, isinf(), and std::isfinite() are all non-standard: and MSVC
does not have them.

It does have _finite() and _isnan(), unless we can find some portable way
of doing things instead.

2. My understand was that you are not intended to store positive infinity,
negative infinity, or NaN in double or float variables - only test
variables for them with functions like isnan, isinf (which unfortunately
are non-standard).

This is because on a particular implementation it is not required that
double/float can store these values.

Is this correct? If so, the use of INFINITY and -INFINITY in GeoTimeInstant
would need to change.
}}}

I looked into this issue, and determined that Glen was correct.  Specifically,
 1. The macros and functions 'INFINITY', 'isinf' and 'isfinite' are part of C99, which of course is not part of the C++ standard.
 1. While the IEEE 754 floating-point standard does indeed define infinity, etc., the C++ standard doesn't require that a C++ implementation use IEEE 754 for floating-point numbers.

For more info, take a look at these links which Google found for me:
 * http://groups.google.com/group/comp.lang.c++/browse_thread/thread/e1cea36d29c4b708/0149a9483297f8e8
 * http://www.thescripts.com/forum/thread165462.html
 * http://gcc.gnu.org/ml/libstdc++/2002-09/msg00038.html


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@1461 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c