Replaced source:/gplates/trunk using source:/gplates/branches/directory-renaming-2007-sep-27#1563 incorporating log:/gplates/branches/directory-renaming-2007-sep-27#1545:1563
This is "[wiki:MergingInSubversion#Replacingonebranchwithanother replacing one branch with another]". The command used was:
{{{
svn merge http://svn.gplates.org/gplates/trunk@1563 http://svn.gplates.org/gplates/branches/directory-renaming-2007-sep-27@1563
}}}
from inside the root directory of a working copy of the trunk.

It may be observed that there haven't been any changesets applied to the trunk between changeset:1545 (which is when this branch was created) and changeset:1563 (the last commit on this branch), so there was no need to sync the trunk to this branch before the merge.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@1564 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c