Replaced source:/gplates/trunk using source:/gplates/branches/fixes-and-cleanups-2007-oct-09#1675 incorporating log:/gplates/branches/fixes-and-cleanups-2007-oct-09#1670:1675
This is "[wiki:MergingInSubversion#Replacingonebranchwithanother replacing one branch with another]". The command used was:
{{{
svn merge http://svn.gplates.org/gplates/trunk@1675 http://svn.gplates.org/gplates/branches/fixes-and-cleanups-2007-oct-09@1675
}}}
from inside the root directory of a working copy of the trunk.

The purpose of these changes was to modify the reconstruction process, so that whenever a feature is encountered which does not possess a reconstruction plate ID, it would still be "reconstructed" with the identity rotation.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@1676 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c