Replaced source:/gplates/trunk using source:/gplates/branches/fixes-and-cleanups-2007-oct-09#1716 incorporating log:/gplates/branches/fixes-and-cleanups-2007-oct-09#1711:1716
This is "[wiki:MergingInSubversion#Replacingonebranchwithanother replacing one branch with another]". The command used was:
{{{
svn merge http://svn.gplates.org/gplates/trunk@1716 http://svn.gplates.org/gplates/branches/fixes-and-cleanups-2007-oct-09@1716
}}}

Note that the only change on the trunk during this time (changeset:1715) was identical to a change applied on this branch (changeset:1714), so we're not losing anything by doing a replace without first sync'ing.
from inside the root directory of a working copy of the trunk.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@1717 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c