Worked around a bug in Qt 4.3.0 which was causing the Edit Feature Properties QTableView to not display any data. See Trolltech Bug 169255: http://trolltech.org/developer/task-tracker/index_html?method=entry&id=169255
Note that on 4.3.0 systems, what seems to be a related bug causes spurious rows to gather on QTableViews as rows are added and deleted from the model. These rows cannot be clicked on, so it should be safe to ignore as a graphical glitch in Qt 4.3.0. They do not appear on Qt 4.3.3.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@2949 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c