Replaced source:/gplates/trunk using source:/gplates/branches/digitisation-2008-feb-27#3359 incorporating log:/gplates/branches/digitisation-2008-feb-27#3306:3359 ([log:/gplates/trunk#3359:3359 No changes on trunk] since last update from trunk).
This is "[wiki:MergingInSubversion#Replacingonebranchwithanother replacing one branch with another]". The command used was:
{{{
svn merge http://svn.gplates.org/gplates/trunk@3359 http://svn.gplates.org/gplates/branches/digitisation-2008-feb-27@3359
}}}
from inside the root directory of a working copy of the trunk.

This updates trunk with the current work on the DigitisationWidget and ExportCoordinatesDialog, as well as an extra field in FeatureSummaryWidget.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@3360 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c