Replaced source:/gplates/trunk using source:/gplates/branches/digitisation-2008-feb-27#3806 incorporating log:/gplates/branches/digitisation-2008-feb-27#3693:3806 ([log:/gplates/trunk#3806:3806 No changes on trunk] since last update from trunk).
This is "[wiki:MergingInSubversion#Replacingonebranchwithanother replacing one branch with another]". The command used was:
{{{
svn merge http://svn.gplates.org/gplates/trunk@3806 http://svn.gplates.org/gplates/branches/digitisation-2008-feb-27@3806
}}}
from inside the root directory of a working copy of the trunk.

This heavily updates the EditGeometryWidget ''(with a hammer)'' so that it can support viewing and editing of all the various GeometryOnSphere derivations we have now.

It also incorporates a number of small fixes in preparation for the 0.9.3 release, chagnes to the ClickGeometry tool (shift-click and empty-click), hides a number of GUI elements we can't support yet (Pole Manipulation tool for one), fixes a crash when deleting geometry from a feature, changes accelerator keys, makes EditTimePeriodWidget more friendly, and adds all manner of FIXMEs.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@3807 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c