Replaced source:/gplates/trunk using source:/gplates/branches/ui-improvements-2008-oct-30#4251 incorporating log:/gplates/branches/ui-improvements-2008-oct-30#4199:4251 ([log:/gplates/trunk#4251:4251 No changes on trunk] since last update from trunk).
This is "[wiki:MergingInSubversion#Replacingonebranchwithanother replacing one branch with another]". The command used was:
{{{
svn merge http://svn.gplates.org/gplates/trunk@4251 http://svn.gplates.org/gplates/branches/ui-improvements-2008-oct-30@4251
}}}
from inside the root directory of a working copy of the trunk.

This merge to trunk adds the ability to create a new FeatureCollection at the end of the d10n step, and adds a Reload button to ManageFeatureCollectionsDialog.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@4252 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c