First set of changes to fix bugs and make users happy:
The main change is I split the Topology workflow into two tools: Build
Topology, Edit Topology.

I removed unecessary widgets from the panels,
and changed Reverse widget
to check box to help clean up bugs with reverse.

As per the user requests to "keep the number of clicks to a minium"
when building and editing topologies,  I did not adjust the edit   
processes to requrire confirmation on a separate tab.

Also to reduce clicks even further, I added a short cut ability on the
Build Topology process to add features to be inserted or removed from the
boundary via Shift Left Click.  This needs to be adjusted for Edit
Topology.




git-svn-id: http://svn.gplates.org/projects/gplates/branches/platepolygon-2008-08-15@4912 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c