Solved a Qt signal/slot '`QObject::connect`' problem, which was resulting in the following run-time complaint (or something similar) : {{{ Object::connect: No such signal GPlatesGui::TopologySectionsContainer::entry_removed(TopologySectionsContainer::size_type deleted_index) Object::connect:  (receiver name: 'BuildTopologyWidget') }}}
The complaint was being written to the terminal/console when the Build Topology tool was activated.

It seemed to be related to namespaces in the signals and slots (I'm trying to cram all the searchable keywords I can think of into this commit message, for the next time we hit this...), but it turns out that this was only part of the issue.

As far as I can tell, the fundamental problem was the presence of the parameter name '`deleted_index`' within the SIGNAL and SLOT macros in the '`QObject::connect`' invocation.

Once that was removed, there were no longer complaints about "No such signal"; instead:
{{{
Object::connect: No such slot GPlatesQtWidgets::BuildTopologyWidget::entry_removed(TopologySectionsContainer::size_type)
Object::connect:  (receiver name: 'BuildTopologyWidget')
}}}
-- because it seems that the signal and slot function parameter types in the SIGNAL and SLOT macros need to match the actual declarations character-for-character, and the slot was declared as:
{{{
void
GPlatesQtWidgets::BuildTopologyWidget::entry_removed(
                GPlatesGui::TopologySectionsContainer::size_type deleted_index)
}}}

And then finally:
{{{
QObject::connect: Incompatible sender/receiver arguments
        GPlatesGui::TopologySectionsContainer::entry_removed(TopologySectionsContainer::size_type) --> GPlatesQtWidgets::BuildTopologyWidget::entry_removed(GPlatesGui::TopologySectionsContainer::size_type)
}}}
because it seems that the signal and slot function parameter types in the SIGNAL and SLOT macros need to match ''each other'' character-for-character too.


git-svn-id: http://svn.gplates.org/projects/gplates/branches/platepolygon-2008-08-15@5778 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c