Added member function '`TopLevelProperty::deep_clone`' for recursive cloning of a top-level property and all the property-values it contains.
In particular, this resulted in the removal of the virtual function '`PropertyValue::clone`' and the addition of the virtual function '`PropertyValue::deep_clone_as_prop_val`', which is invoked by '`TopLevelProperty::deep_clone`'.

Every derivation of PropertyValue (but not PropertyValue itself) additionally gained two more non-virtual functions:
 * `clone`
 * `deep_clone`

Both of these non-virtual functions return a `non_null_ptr_type` to the '''derived''' type, rather than to PropertyValue.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@7833 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c