(trunk) Fixed bug where a new revision ID would be generated for features on load from file.
The cause of the problem is that previously, a new revision ID was generated whenever a property in a feature was added, removed or changed. However, the file readers pretty much work like this: create the feature, and then create the properties and put them into the feature. When those properties were put into the feature, the feature would get a new revision ID, even though the user had not modified the feature.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@8063 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c