(trunk) Roll back bug Fix: #232: Remove duplicate file extension when exporting velocities. See below Mark's comments. I agree with him. So, roll back the fix for now.
hi all,

This is actually not a bug;
The total file name has several components.

This is the full file name:

"velocity_colat+lon_at_000_140.00Ma_on_mesh-9.mesh.0.gpml.gpml"

This part says what the contents are 
"velocity_colat+lon"


This part shows the index number, 000, and the age 140
"_at_000_140.00Ma"


This part shows which original mesh file the data is computed on:
"on_mesh-9.mesh.0.gpml"

And this part shows the file format of the velocity file.
".gpml"


So the seemingly duplicate file extension is actually there on purpose.

At some point we might read in other mesh formats,
".abc" or  ".Terra" or ".whatever" ,
and so the apparent duplicate extension would show up...


Sabin, as for the script not working, I made some adjustments to
convert_meshes_gpml_to_citcoms.py on June 10, so you might need to get a 
new copy.

Give that a try and if things are still not working we can talk tonight.

Mark


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@8915 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c