(trunk) Additional improvements to reconstruction iterator related code, to further protect against crashes as per ticket:261.
get_next_valid_reconstruction_geometry_collection_iterator in ConstReconstructionGeometryIterator now takes a non-const reference to the reconstruction tree map iterator from which to start looking, instead of a temporary - see comments above function.
This change actually negates the need for the previous fix in changeset:9653, but I've left that in as double protection against Bad Things happening - this is explained in the comments in operator== for ConstReconstructionGeometryIterator.
Also added size() and empty() functions to ReconstructionGeometryCollection, for good measure.

git-svn-id: http://svn.gplates.org/projects/gplates/trunk@9660 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c