(trunk) Changed {{{GPLATES_PUBLIC_RELEASE}}} back to false (was changed to true in changeset:10128).
Modified CMake scripts so that the {{{profilegplates}}} build environment is created regardless of the value of {{{GPLATES_PUBLIC_RELEASE}}}. We require CMake 2.6 now anyway, so the problems with 2.4 should have gone away. Also, when {{{GPLATES_PUBLIC_RELEASE}}} was set to true, CMake 2.6 failed to generate, complaining that the {{{CMAKE_EXE_LINKER_FLAGS_PROFILEGPLATES}}} variable was not set.

git-svn-id: http://svn.gplates.org/projects/gplates/trunk@10133 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c