Replaced source:/gplates/trunk using source:/gplates/branches/configuration-2011-Aug-17@12215 incorporating log:/gplates/branches/configuration-2011-Aug-17@12125-12215 ([log:/gplates/trunk@12214-12215 No changes on trunk] since last update from trunk).
This is "[wiki:MergingInSubversion#Replacingonebranchwithanother replacing one branch with another]". The command used was:
{{{
svn merge https://svn-test.gplates.org/gplates/trunk@12215 https://svn-test.gplates.org/gplates/branches/configuration-2011-Aug-17@12215
}}}
from inside the root directory of a working copy of the trunk.

This adds a basic GPlatesUtils::ConfigBundle class that can be used in conjunction with, or separately to, UserPreferences.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@12216 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c