Replaced source:/gplates/trunk using source:/gplates/branches/configuration-2011-Aug-17@12404 incorporating log:/gplates/branches/configuration-2011-Aug-17@12217-12404 ([log:/gplates/trunk@12404-12404 No changes on trunk] since last update from trunk).
This is "[wiki:MergingInSubversion#Replacingonebranchwithanother replacing one branch with another]". The command used was:
{{{
svn merge https://svn-test.gplates.org/gplates/trunk@12404 https://svn-test.gplates.org/gplates/branches/configuration-2011-Aug-17@12404
}}}
from inside the root directory of a working copy of the trunk.

This adds the "Advanced" UI for UserPreferences, and the ability to make a QTableView linked to a ConfigBundle. It's editable, but the delegates aren't finished yet.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@12405 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c