Replaced source:/gplates/trunk using source:/gplates/branches/data-mining-python-api-2011-Oct-25@13670 incorporating log:/gplates/branches/data-mining-python-api-2011-Oct-25@13616-13670 ([log:/gplates/trunk@13644-13670 no change on trunk] since last update from trunk).
This is "[wiki:MergingInSubversion#Replacingonebranchwithanother replacing one branch with another]". The command used was:
{{{
svn merge http://svn.gplates.org/gplates/trunk http://svn.gplates.org/gplates/branches/data-mining-python-api-2011-Oct-25
}}}
from inside the root directory of a working copy of the trunk.

This adds a basic GPlatesUtils::ConfigBundle class that can be used in conjunction with, or separately to, UserPreferences.


git-svn-id: http://svn.gplates.org/projects/gplates/trunk@13671 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c