Add memory management of fragments in scene tile.
The algorithm follows that of:
"Lock-Free Linked Lists Using Compare-and-Swap"  Valois 1995, with
corrections covered in: "Correction of a Memory Management Method
for Lock-Free Data Structures" Scott 1996.
It is a reference-counting scheme that tracks all shared and private
pointers to fragments, and reclaims a fragment (for re-use) when there
are no longer any references to it, and also solves the ABA problem
(https://en.wikipedia.org/wiki/ABA_problem).