If a feature has no default geometry then return non-default geometry.
This is more useful when you are expecting a single geometry but don’t
know which property name it’s under (because it might've been imported
into a non-default geometry unexpectedly). Previously only the default
geometry property name was queried.
The same applies to the default topological geometry in a feature.