Fix long-standing net rotation bug when some plates are stationary.
Topologies that are stationary now contribute area to the normalization
weight (whereas previously did not). This increases the normalization
denominator and hence decreases the net rotation. But if no plates are
stationary then the results should be the same as before.
Also, recent changes fixed a Vector3D normalization exception during net
rotation export when stationary plates are present.